library_system/reservation_system/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
django-cors-headers
psycopg2
pyjwt
cryptography